Crate sgx_urts

Modules

asyncio

env

event

fd

file

mem

net

pipe

process

signal

socket

sys

thread

time

Structs

SgxEnclave

Functions

rsgx_create_enclave
Loads the enclave using its file name and initializes it using a launch token.

rsgx_create_enclave_from_buffer_ex

rsgx_create_enclave_with_workers
Loads the enclave using its file name and initializes it using a launch token and certain number of uworkers and tworkers.

rsgx_create_encrypted_enclave

rsgx_destroy_enclave
The function destroys an enclave and frees its associated resources.

rsgx_get_target_info